test-arena2/pyproject.toml
[project]
name = "test-arena2"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "pikamonvvs", email = "pikamonvvs@naver.com"},
]
dependencies = [
    "loguru>=0.7.2",
    "anyio>=4.4.0",
    "ffmpeg-python>=0.2.0",
    "httpx>=0.27.0",
    "streamlink>=6.8.3",
    "httpx-socks>=0.9.1",
    "aiohttp>=3.9.5",
    "types-requests>=2.32.0.20240712",
    "h2>=4.1.0",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
